model_serving/requirements.txt

fastapi
uvicorn
pandas
scikit-learn>=1.6.0
joblib
google-cloud-storage
